pyproject.toml
[tool.poetry]
name = "grammar-guard"
version = "0.1.0"
description = "A lightweight, runtime-agnostic constrained decoding layer for reliable JSON output from LLMs"
authors = ["Meshari Alaeena <mesharialaeena@outlook.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/malaeena/grammar-guard"
repository = "https://github.com/malaeena/grammar-guard"
keywords = ["llm", "json", "constrained-decoding", "structured-output", "schema-validation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [{include = "grammar_guard"}]

[tool.poetry.dependencies]
python = "^3.9"

# Core schema handling
pydantic = "^2.5"
jsonschema = "^4.20"

# FSM and regex for constrained decoding
interegular = "^0.3"

# Model backends
transformers = "^4.36"
torch = "^2.1"
accelerate = "^0.25"
llama-cpp-python = "^0.2.27"

# CLI and terminal UI
typer = {extras = ["all"], version = "^0.9"}
rich = "^13.7"
click = "^8.1"

# Utilities
loguru = "^0.7"
typing-extensions = "^4.9"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-benchmark = "^4.0"
hypothesis = "^6.92"

# Code quality
black = "^23.12"
ruff = "^0.1"
mypy = "^1.7"
pre-commit = "^3.6"

# Documentation
mkdocs = "^1.5"
mkdocs-material = "^9.5"

[tool.poetry.scripts]
grammar-guard = "grammar_guard.cli.main:cli"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "interegular.*",
    "llama_cpp.*",
    "transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=grammar_guard",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (require models)",
    "e2e: End-to-end tests (full scenarios)",
    "benchmark: Benchmark tests (performance measurement)",
    "slow: Slow tests (skip with -m 'not slow')",
]

[tool.coverage.run]
source = ["grammar_guard"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
